added new property UTF8Text
A property not in the official doucmentation 0x0709 is according to a forums post a duplication of a normal text property but utf8 encoded. This does not appear in cdxml files.
Added code to specifcially handle this property.  With this addition a file with a molecule and styled text can be read in and written back out as cdx being 100% identically on binary level.