fixed CDXPoint3D coordinate order
The official specification says the coordiante order is z,y,x for cdx and x,y,z for cdxml.
However testing with sample files in ChemDraw DataDir shows that doing it that way leads to corrupt cdx.
Using also order x,y,z for cdx and the corruption is gone and files look correct when viewd in ChemDraw.
Therefore the official spec is either wrong or was changed at some point.