basic charset support when reading cdx
Each style run can have a different font and each font can have a different charset. A Text object in cdx however encodes the entire text in 1 block so cdx/cdxml aren't 100% compatible in this regard as a Text object or <t> element is in cdx expected to be of the same encoding for all styles.
Therefore the handling might need to be changed/simplified in the future